refactor(currencies): promote drawers/dialogs from the page to v2/components
将来的な共有可能性（設計システムの action-sheet / destructive-confirm パターン）を
見越し、ページ専用扱いをやめて feature の components/ 配下へ昇格。

- currency-actions-drawer / transaction-actions-drawer / delete-currency-dialog /
  delete-transaction-dialog を pages/currency-detail-page/ から
  v2/components/ へ移設（中身は currency 固有のまま、置き場所のみ変更）。
- currency-detail-page.tsx の該当 import を相対 → 絶対(components) に更新。
- TopBar と詳細ページ内容コンポーネント（balance-hero / description /
  detail-skeleton / transaction-form / transaction-list）はページ固有のため据え置き。

これらは props のみの純表示コンポーネントで hooks 不使用。挙動・DOM 不変。
描画/配線の検証は引き続き currency-detail-page.test.tsx が実体レンダリングでカバー。
