+++ apps/web/src/features/currencies/v2/components/delete-currency-dialog/delete-currency-dialog.tsx
