+++ apps/web/src/features/currencies/v2/components/delete-transaction-dialog/index.ts
