+++ apps/web/src/features/currencies/v2/pages/currency-detail-page/currency-detail-page.tsx
@@ -1,16 +1,16 @@
 import { IconPlus, IconStar, IconStarFilled } from "@tabler/icons-react";
+import { CurrencyActionsDrawer } from "@/features/currencies/v2/components/currency-actions-drawer";
 import { CurrencyFormV2 } from "@/features/currencies/v2/components/currency-form";
+import { DeleteCurrencyDialog } from "@/features/currencies/v2/components/delete-currency-dialog";
+import { DeleteTransactionDialog } from "@/features/currencies/v2/components/delete-transaction-dialog";
+import { TransactionActionsDrawer } from "@/features/currencies/v2/components/transaction-actions-drawer";
 import { FormSheet } from "@/shared/components/form-sheet";
 import { PageHeader } from "@/shared/components/page-header";
 import { Button } from "@/shared/components/ui/button";
-import { CurrencyActionsDrawer } from "./currency-actions-drawer";
 import { CurrencyBalanceHero } from "./currency-balance-hero";
 import { CurrencyDescription } from "./currency-description";
 import { CurrencyDetailSkeleton } from "./currency-detail-skeleton";
-import { DeleteCurrencyDialog } from "./delete-currency-dialog";
-import { DeleteTransactionDialog } from "./delete-transaction-dialog";
 import { TopBar } from "./top-bar";
-import { TransactionActionsDrawer } from "./transaction-actions-drawer";
 import { TransactionFormV2 } from "./transaction-form";
 import { TransactionListV2 } from "./transaction-list";
 import { useCurrencyDetailPage } from "./use-currency-detail-page";
